refactor: рефактор страницы Calendar по замечаниям
