+++ src/pages/Books/index.js
@@ -2,7 +2,7 @@ import BasePage from '../../layout/BasePage';
 import CardBook from '../../components/Cards/CardBook/CardBook';
 import AnimatedPageContainer from '../../components/AnimatedPageContainer/AnimatedPageContainer';
 import TagsList from '../../components/TagsList/TagsList';
-import { Loader, Paginate, TitleH1 } from '../../components/utils/index';
+import { Loader, Paginate, TitleH1 } from '../../components/utils';
 
 export {
   BasePage,
