+++ src/pages/Calendar/index.js
@@ -2,7 +2,7 @@ import BasePage from '../../layout/BasePage';
 import CardCalendar from '../../components/Cards/CardCalendar/CardCalendar';
 import AnimatedPageContainer from '../../components/AnimatedPageContainer/AnimatedPageContainer';
 import TagsList from '../../components/TagsList/TagsList';
-import { TitleH1, Loader, Paginate } from '../../components/utils/index';
+import { Loader, Paginate, TitleH1 } from '../../components/utils';
 
 export {
   BasePage,
